requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-cors-headers==3.14.0
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
psycopg2==2.9.5
pycodestyle==2.10.0
python-decouple==3.8
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
tzdata==2023.3
